refactor: extract page sections and shared link primitives
app/page.tsx was a 314-line file mixing build-time data fetching with five
inline presentational sub-components. Decompose it so page.tsx is a thin
composition root (~70 lines) and each section is its own component, matching
the rest of the tree:

- components/{Hero,SectionHeader,WritingSection,RecommendationCard,
  RecommendationsSection}.tsx
- components/ExternalLink.tsx: the target=_blank + rel=noopener noreferrer
  idiom (repeated ~10x) as one primitive; adopted in page sections, Footer,
  ActivityFeed, and the chat page.
- components/AsideLink.tsx: the SectionHeader aside link with the enlarged
  tap-target, copy-pasted 3x, as one component.

Behavior and markup are unchanged.
