+++ app/page.tsx
@@ -1,253 +1,15 @@
-import Link from "next/link";
 import Reveal from "@/components/Reveal";
 import ActivityFeed from "@/components/ActivityFeed";
 import GithubHighlights from "@/components/GithubHighlights";
-import { getSubstackPosts, type SubstackPost } from "@/lib/sources/substack/substack";
+import Hero from "@/components/Hero";
+import SectionHeader from "@/components/SectionHeader";
+import AsideLink from "@/components/AsideLink";
+import WritingSection from "@/components/WritingSection";
+import RecommendationsSection from "@/components/RecommendationsSection";
+import { getSubstackPosts } from "@/lib/sources/substack/substack";
 import { getRepoHighlights } from "@/lib/sources/github/repos";
 import { getGithubActivity } from "@/lib/sources/github/activity";
-import { recommendations, type Recommendation } from "@/data/linkedin";
-import { portrait, recommenderAvatars } from "@/data/avatars";
 import { site } from "@/lib/site";
-import { formatDate } from "@/lib/format";
-
-function SectionHeader({
-  kicker,
-  title,
-  aside,
-}: {
-  kicker: string;
-  title: React.ReactNode;
-  aside?: React.ReactNode;
-}) {
-  return (
-    <div className="flex flex-wrap items-end justify-between gap-3">
-      <div>
-        <p className="kicker">{kicker}</p>
-        <h2 className="display mt-2 text-3xl sm:text-4xl">{title}</h2>
-      </div>
-      {aside && <div className="pb-1 font-mono text-xs text-muted">{aside}</div>}
-    </div>
-  );
-}
-
-function Hero() {
-  return (
-    <section className="mx-auto grid max-w-5xl items-center gap-10 px-5 pb-16 pt-14 sm:pt-20 md:grid-cols-[1.2fr_0.8fr]">
-      <div>
-        <p className="kicker">Hello, I’m</p>
-        <h1 className="display mt-3 text-5xl sm:text-6xl md:text-7xl">
-          Andrew <em>Carpenter</em>
-        </h1>
-        <p className="mt-6 max-w-xl text-lg leading-relaxed text-soft">
-          {site.tagline} This is the running record — live from{" "}
-          <a href={site.githubUrl} className="font-medium text-accent hover:text-accent-deep" target="_blank" rel="noopener noreferrer">
-            GitHub
-          </a>
-          ,{" "}
-          <a href={site.substackUrl} className="font-medium text-accent hover:text-accent-deep" target="_blank" rel="noopener noreferrer">
-            Substack
-          </a>
-          , and{" "}
-          <a href={site.linkedinUrl} className="font-medium text-accent hover:text-accent-deep" target="_blank" rel="noopener noreferrer">
-            LinkedIn
-          </a>
-          .
-        </p>
-        <div className="mt-8 flex flex-wrap items-center gap-3">
-          <Link href="/story/" className="btn btn-primary">
-            Read my story
-          </Link>
-          <Link href="/philosophy/" className="btn btn-secondary">
-            What I believe
-          </Link>
-          <Link
-            href="/chat/"
-            className="px-2 py-2.5 text-sm font-medium text-muted transition-colors hover:text-accent"
-          >
-            or just ask me →
-          </Link>
-        </div>
-      </div>
-
-      <div className="relative mx-auto w-56 sm:w-64 md:w-full md:max-w-[280px]">
-        <div aria-hidden className="absolute inset-0 translate-x-3 translate-y-3 rounded-2xl border border-line-strong bg-paper" />
-        <div className="card relative overflow-hidden rounded-2xl p-2">
-          {/* eslint-disable-next-line @next/next/no-img-element */}
-          <img
-            src={portrait.src}
-            alt={`Portrait of ${site.name}`}
-            width={320}
-            height={320}
-            fetchPriority="high"
-            className="aspect-square w-full rounded-lg bg-paper object-cover outline -outline-offset-1 outline-black/10"
-          />
-          <p className="px-2 pb-1 pt-2 text-center font-mono text-[0.65rem] uppercase tracking-widest text-faint">
-            Est. somewhere — still shipping
-          </p>
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function WritingSection({ posts }: { posts: SubstackPost[] }) {
-  return (
-    <section className="mx-auto max-w-5xl px-5 pt-20" id="writing">
-      <Reveal>
-        <SectionHeader
-          kicker="Writing"
-          title={
-            <>
-              From the <em>Substack</em>
-            </>
-          }
-          aside={
-            <a
-              href={site.substackUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="relative inline-block transition-colors before:absolute before:-inset-x-1 before:-inset-y-3 before:content-[''] hover:text-accent"
-            >
-              {site.substackHandle} ↗
-            </a>
-          }
-        />
-      </Reveal>
-
-      {posts.length > 0 ? (
-        <div className="mt-8 grid gap-4 md:grid-cols-3">
-          {posts.slice(0, 3).map((post, i) => (
-            <Reveal key={post.link} delay={i * 90}>
-              <a href={post.link} target="_blank" rel="noopener noreferrer" className="card group flex h-full flex-col px-6 py-6">
-                <p className="font-mono text-xs text-muted">{formatDate(post.date)}</p>
-                <h3 className="display mt-3 text-xl leading-snug group-hover:text-accent transition-colors">
-                  {post.title}
-                </h3>
-                {post.excerpt && (
-                  <p className="mt-3 flex-1 text-sm leading-relaxed text-muted">{post.excerpt}</p>
-                )}
-                <p className="mt-4 text-sm font-medium text-accent">Read on Substack →</p>
-              </a>
-            </Reveal>
-          ))}
-        </div>
-      ) : (
-        <Reveal>
-          <a
-            href={site.substackUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="card mt-8 block px-6 py-10 text-center"
-          >
-            <p className="display text-xl">
-              Fresh posts land here on every deploy<span className="text-accent">.</span>
-            </p>
-            <p className="mt-2 text-sm text-muted">
-              In the meantime, read everything at {site.substackUrl.replace("https://", "")} ↗
-            </p>
-          </a>
-        </Reveal>
-      )}
-    </section>
-  );
-}
-
-function RecommendationCard({ rec }: { rec: Recommendation }) {
-  const avatar = rec.profileUrl ? recommenderAvatars[rec.profileUrl] : undefined;
-  const inner = (
-    <>
-      <span aria-hidden className="display text-5xl leading-none text-accent">
-        “
-      </span>
-      <blockquote className="mt-1 whitespace-pre-line text-sm leading-relaxed text-soft">
-        {rec.quote}
-      </blockquote>
-      <figcaption className="mt-5 flex items-center gap-3 border-t border-line pt-4">
-        {avatar && (
-          /* eslint-disable-next-line @next/next/no-img-element */
-          <img
-            src={avatar.src}
-            alt=""
-            width={40}
-            height={40}
-            loading="lazy"
-            decoding="async"
-            className="h-10 w-10 shrink-0 rounded-full object-cover outline -outline-offset-1 outline-black/10"
-          />
-        )}
-        <span className="min-w-0">
-          <span className="flex flex-wrap items-center gap-x-2 text-sm font-semibold text-ink">
-            <span className="group-hover:text-accent transition-colors">{rec.name}</span>
-            {rec.profileUrl && (
-              <span aria-hidden className="font-mono text-xs font-normal text-faint group-hover:text-accent">
-                ↗
-              </span>
-            )}
-            {rec.sample && <span className="chip font-normal">sample</span>}
-          </span>
-          <span className="mt-0.5 block text-xs text-muted">
-            {rec.title}
-            {rec.relationship ? ` · ${rec.relationship}` : ""}
-          </span>
-        </span>
-      </figcaption>
-    </>
-  );
-
-  const className = "card group block break-inside-avoid px-6 py-6";
-  return rec.profileUrl ? (
-    <a
-      href={rec.profileUrl}
-      target="_blank"
-      rel="noopener noreferrer"
-      className={className}
-      aria-label={`View ${rec.name}’s LinkedIn profile`}
-    >
-      {inner}
-    </a>
-  ) : (
-    <figure className={className}>{inner}</figure>
-  );
-}
-
-function RecommendationsSection() {
-  return (
-    <section className="mx-auto max-w-5xl px-5 pt-20" id="kind-words">
-      <Reveal>
-        <SectionHeader
-          kicker="Kind words"
-          title={
-            <>
-              What colleagues <em>say</em>
-            </>
-          }
-          aside={
-            <a
-              href={site.linkedinUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="relative inline-block transition-colors before:absolute before:-inset-x-1 before:-inset-y-3 before:content-[''] hover:text-accent"
-            >
-              recommendations via LinkedIn ↗
-            </a>
-          }
-        />
-      </Reveal>
-
-      <div className="mt-8 gap-4 sm:columns-2 lg:columns-3">
-        {recommendations.map((rec, i) => (
-          <Reveal
-            key={rec.name}
-            delay={Math.min(i, 3) * 80}
-            className="reveal mb-4 break-inside-avoid"
-          >
-            <RecommendationCard rec={rec} />
-          </Reveal>
-        ))}
-      </div>
-    </section>
-  );
-}
 
 export default async function Home() {
   const [posts, repos, githubActivity] = await Promise.all([
@@ -288,16 +50,7 @@ export default async function Home() {
                 Things I’ve <em>built</em>
               </>
             }
-            aside={
-              <a
-                href={site.githubUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="relative inline-block transition-colors before:absolute before:-inset-x-1 before:-inset-y-3 before:content-[''] hover:text-accent"
-              >
-                github.com/{site.githubUsername} ↗
-              </a>
-            }
+            aside={<AsideLink href={site.githubUrl}>github.com/{site.githubUsername} ↗</AsideLink>}
           />
         </Reveal>
         <Reveal delay={120}>
