+++ components/ExternalLink.tsx
@@ -1,0 +1,19 @@
+import type { AnchorHTMLAttributes, ReactNode } from "react";
+
+type ExternalLinkProps = AnchorHTMLAttributes<HTMLAnchorElement> & {
+  href: string;
+  children: ReactNode;
+};
+
+/**
+ * Anchor to an off-site URL with the standard new-tab safety rel applied once,
+ * in one place. Any other anchor attribute (className, aria-label, …) passes
+ * straight through.
+ */
+export default function ExternalLink({ href, children, ...props }: ExternalLinkProps) {
+  return (
+    <a href={href} target="_blank" rel="noopener noreferrer" {...props}>
+      {children}
+    </a>
+  );
+}
